Update metrics and activation function; rename lecture files
The 'mse' metric in the Sequential API example has been replaced with 'accuracy'. Additionally, the activation function in the advanced lecture code was changed from 'relu' to 'PReLU'. A significant number of lecture files in the advanced section were also renamed to more descriptive, numbered titles for better organization and understanding.