Add Seq2Seq model script for neural machine translation
This commit introduces an initial script to build a Seq2Seq model for neural machine translation (NMT) using PyTorch. The code includes functions to tokenize sentences and build vocabularies, classes for the encoder and decoder modules, and a Seq2Seq model that leverages these modules. It also contains dummy sentences for demonstrating tokenization and a training loop to demonstrate model training.